Add files via upload
In this notebook a version of the SSP4-3.4 scenario is run in which temperatures are capped at 1.5˚C in order to model a "shave-the-peak" solar geoengineering scenario. This is compared to SSP4-3.4 in order to roughly quantify the effects of solar geoengineering on atmospheric CO2 through its effects on the carbon cycle.